azure/cosmosdb/sdk/python/tests/test_mockup_data_store.py
import pytest


def test_importable():
    import src.mockup_data_store  # noqa: F401
    from src.mockup_data_store import MockupDataStore  # noqa: F401


def test_init():
    from src.mockup_data_store import MockupDataStore
    data_store = MockupDataStore()
    assert data_store is not None

@pytest.mark.parametrize("item", [
    {'key': 'value'},
])
def test_insert_item(item):
    from src.mockup_data_store import MockupDataStore
    data_store = MockupDataStore()
    data_store.insert_item(item)

@pytest.mark.parametrize("max_item_count", [
    1, 10
])
def test_read_items(max_item_count):
    from src.mockup_data_store import MockupDataStore
    data_store = MockupDataStore()
    items = data_store.read_items(max_item_count=max_item_count)
    assert items is not None
    assert isinstance(items, list), f"Expected list, but got {type(items)}"
    assert len(items) <= max_item_count

@pytest.mark.parametrize("field_name_value", [
    {'key': 'value'}
])
def test_query_items(field_name_value):
    from src.mockup_data_store import MockupDataStore
    data_store = MockupDataStore()
    responses = data_store.query_items(field_name_value=field_name_value)
    assert responses is not None
    assert isinstance(responses, list), \
        f"Expected list, but got {type(responses)}"
    for response in responses:
        assert isinstance(response, dict)
